% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMaxt5.R
\docType{package}
\name{AMaxt5}
\alias{AMaxt5}
\title{AMax-T test package}
\description{
This package find test statistica value and critical value of asymptotic
Max-T test for testing
equality of means against ordered alternative in one-way ANOVA model
with heteroscedasticity when number of groups is 5.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
